Bug fixed for showing error if non-owner of the post tries to delete the post. Code changes made in ImageController and ownership is validated before processing the request.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -150,9 +150,19 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        User currentUser= (User) session.getAttribute("loggeduser");
+        if(currentUser.getId()==(imageService.getImage(imageId)).getUser().getId())
+        {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("deleteError",error);
+            model.addAttribute("image", imageService.getImage(imageId));
+            return  "images/image";
+        }
     }
